Camera: Fix misuse of ASSERT() within a function
ASSERT() is effectively a `return` statement. So if
ASSERT() is used in a function called by a test case, it won't
stop the test case all together, but rather returns from the function
and continue to run the test case to completion.

Using ASSERT_NO_FATAL_FAILURE makes sure the rest of the test
case is skipped.

Also fix a wrong check in openEmptyDeviceSession.

Flag: TEST_ONLY
Test: run VtsAidlHalCameraProvider_TargetTest
Bug: 399939768
Change-Id: Id33e22b1e8eb76a2f7f02dcf7308c86c55aad68d