VTS: Update dataspace and readback buffer support after setColorMode
[Why]
Dataspace can get updated after calls to
ComposerClientWrapper::setColorMode so we can't cache it at setup.
Dataspace is used to check for readback support, so we have to check for
the support after we get the current dataspace.

[How]
Create a helper function that retrieves the dataspace to use in the
subtest, and check if readback is supported given the dataspace.
Call this function after every call to setColorMode when the dataspace
can change.

Bug: b/395405757
Test: atest VtsHalGraphicsComposer3_TargetTest
Flag: TEST_ONLY
Change-Id: I02d7e0aa8915517664b80efec7a72bf7f8773962