Add missing RustDerive attributes to radio AIDL interfaces
In a previous CL, I've added the partialEq, Eq and Clone rust derives to
most parceleables in the radio AIDL files. I missed a few, so this CL
fixes that. These rust derives are needed to be able to clone these
data types, and to perform comparisons. All parceleables were included,
except 'ims/media/LocalEndPoint.aidl' since it contains data types that
are non clonable.

Bug: 393444294
Test: m android.hardware.radio.sim-update-api android.hardware.radio.network-update-api && m
Test: use following command to find files that need to be updated
Test: "grep Rust $(grep -R  parcelable . -l) -L | sort"
Change-Id: I1c8e0f9c3746c41ce9a6c113c1fd5682c54ae262