Get the latest hotplugs
[Why]
With VKMS, we want to test if the hwc indeed receives an hdd from the
kernel and send a hotplug event to its client when the connector status
changes.

[How]
Create a list of hotplugs that gets updated on every hotplug event
callback from the hwc. Save with its ID and connection status.
Once it's read, clear it, so the client is able to always check for the
latest since it was last read, or clear to listen to the newest one.

Bug: b/398831713, b/352790467
Test: atest hotplugs_test
Flag: TEST_ONLY
Change-Id: If90359848816e9c974858952ca117987ef1b763c